OneDrive/デスクトップ/we_planet_backend_login_mypage/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pymysql==1.1.1
alembic==1.13.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-jose[cryptography]
passlib[bcrypt]